requirements.txt
flask
fastapi
pillow
uvicorn
python-multipart
requests